엔티티 PK 타입 Long→Integer 수정 (DB int 타입 일치)
AccommodationCategory, Company, Region의 PK가 DB에서 int인데
엔티티가 Long(bigint)으로 되어있어 FK 타입 불일치 DDL 에러 발생.
관련 Repository ID 타입도 동일하게 수정.

+++ src/main/java/com/honeyrest/honeyrest_host/entity/Company.java
@@ -20,7 +20,7 @@ public class Company extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "company_id")
-    private Long companyId; // 업체 고유 식별자
+    private Integer companyId; // 업체 고유 식별자
 
     @Column(name = "name", nullable = false, length = 200)
     private String name; // 업체명
