+++ src/main/java/com/honeyrest/honeyrest_host/repositoryOwner/ORegionRepository.java
@@ -3,5 +3,5 @@ package com.honeyrest.honeyrest_host.repositoryOwner;
 import com.honeyrest.honeyrest_host.entity.Region;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ORegionRepository extends JpaRepository<Region, Long> {
+public interface ORegionRepository extends JpaRepository<Region, Integer> {
 }
